// types/getPlannerDto.ts
/**
 * Generated by orval v6.10.3 🍺
 * Do not edit manually.
 * culinaryApp
 * OpenAPI spec version: 1.0
 */
import type { GetPlannerRecipeDto } from './getPlannerRecipeDto';
import type { GetProductFromPlannerDto } from './getProductFromPlannerDto';

export interface GetPlannerDto {
  id?: number;
  date?: string;
  plannerRecipes?: GetPlannerRecipeDto[] | null;
  products?: GetProductFromPlannerDto[] | null;
}